seprate the txparams type into a new module. place constructors in right place. change default behavior of initializing TX params, to find a good upstream from the list of upstream_nodes.
