+++ ol/txs/src/commands/demo_cmd.rs
@@ -6,7 +6,11 @@ use abscissa_core::{Command, Options, Runnable};
 use diem_json_rpc_types::views::TransactionView;
 use diem_transaction_builder::stdlib as transaction_builder;
 use ol_types::config::TxType;
-use crate::{entrypoint, submit_tx::{TxError, TxParams, maybe_submit, tx_params_wrapper}};
+use crate::{
+  entrypoint, 
+  submit_tx::{TxError, maybe_submit, tx_params_wrapper},
+  tx_params::TxParams,
+};
 use std::process::exit;
 use std::path::PathBuf;
