+++ ol/txs/src/lib.rs
@@ -23,6 +23,7 @@ pub mod error;
 pub mod prelude;
 pub mod submit_tx;
 pub mod save_tx;
+pub mod tx_params;
 pub mod sign_tx;
 pub mod relay;
 pub mod epoch;
